outputstream problem is sovled, but needs to be oprimized

+++ core/src/main/scala/org/apache/spark/shuffle/IndexShuffleBlockResolver.scala
@@ -61,17 +61,20 @@ private[spark] class IndexShuffleBlockResolver(
    * Remove data file and index file that contain the output data from one map.
    * */
   def removeDataByMap(shuffleId: Int, mapId: Int): Unit = {
-    var file = getDataFile(shuffleId, mapId)
-    if (file.exists()) {
-      if (!file.delete()) {
-        logWarning(s"Error deleting data ${file.getPath()}")
+    val result = conf.getBoolean("spark.shuffle.deleteshuffledata", true)
+    if (result) {
+      var file = getDataFile(shuffleId, mapId)
+      if (file.exists()) {
+        if (!file.delete()) {
+          logWarning(s"Error deleting data ${file.getPath()}")
+        }
       }
-    }
 
-    file = getIndexFile(shuffleId, mapId)
-    if (file.exists()) {
-      if (!file.delete()) {
-        logWarning(s"Error deleting index ${file.getPath()}")
+      file = getIndexFile(shuffleId, mapId)
+      if (file.exists()) {
+        if (!file.delete()) {
+          logWarning(s"Error deleting index ${file.getPath()}")
+        }
       }
     }
   }
